requirements.txt
streamlit
torch
torchvision
python-dotenv
langchain-groq